Initial commit
This commit contains the Fusion Pipeline code at the state in which it existed
at the time I was told to commit it to a git repository. There were a few fixes
I implemented, mainly to make the code compile under Java 1.6. Some files that
were not in use were omitted, in particular old versions of plugins that have
since been rolled into the Fusion Pipeline plugin.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>