Tri-view processor
This processor allows a user to draw a ROI around the sample in three
orthogonal views. From these ROIs, a hull is determined, which is then
used to first reject beads within the shape (as these are false
detections, picking up the sample instead of an actual bead), then black
out the image data outside the hull (as that data is not desired after
reconstruction).

Warning: This processor currently irrecoverably alters the original data.
It is recommended that the originals be backed up first. (Conveniently,
the OME-TIFF metadata remover backs up images.)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>